Moves seeding completely to generate.js
