+++ db/index.js
@@ -9,26 +9,6 @@ const connection = mysql.createConnection({
 
 connection.connect();
 
-const seedData = () => {
-  connection.query(
-    `TRUNCATE TABLE photos`, (error, results, fields) => {
-      if (error) throw error;
-      console.log(results);
-    },
-  );
-  connection.query(
-    `LOAD DATA LOCAL INFILE '/tmp/shoes.csv'
-    INTO TABLE photos
-    FIELDS TERMINATED BY ','
-    LINES TERMINATED BY '\n' 
-    (photo,user,likes,posted_on,product)`, (error, results, fields) => {
-      if (error) throw error;
-      console.log(results);
-    },
-  );
-  connection.end();
-};
-
 const getGramInfo = (id, cb) => {
   connection.query('select * from photos where product = ?', [id], (error, results, fields) => {
     if (error) {
@@ -37,10 +17,9 @@ const getGramInfo = (id, cb) => {
       cb(null, results);
     }
   });
-  connection.end();
 };
 
 module.exports = {
   getGramInfo,
-  seedData,
+  connection,
 };
