+++ generate.js
@@ -1,6 +1,6 @@
 const fs = require('fs');
 const faker = require('faker');
-const { seedData } = require('./db');
+const { connection } = require('./db');
 
 
 const out = fs.createWriteStream('/tmp/shoes.csv');
@@ -10,5 +10,26 @@ const records = Array(500).fill().map((e, i) => {
 records.forEach((i) => {
   out.write(`${i}\n`);
 });
+
+
+const seedData = () => {
+  connection.query(
+    'TRUNCATE TABLE photos', (error, results, fields) => {
+      if (error) throw error;
+      console.log(results);
+    },
+  );
+  connection.query(
+    `LOAD DATA LOCAL INFILE '/tmp/shoes.csv'
+    INTO TABLE photos
+    FIELDS TERMINATED BY ','
+    LINES TERMINATED BY '\n' 
+    (photo,user,likes,posted_on,product)`, (error, results, fields) => {
+      if (error) throw error;
+      console.log(results);
+    },
+  );
+  connection.end();
+};
 
 seedData();
